# fix: optimize reading byte-aligned types from a byte-aligned buffer