--- a/src/buf/byte_order.rs
+++ b/src/buf/byte_order.rs
@@ -8,7 +8,14 @@ use crate::prelude::*;
 /// Endian).
 pub trait ByteOrder {
     fn load<O: BitStore, U: Integral>(src: &BitSlice<O>) -> U;
+    fn load_u16(src: &[u8]) -> u16;
+    fn load_u24(src: &[u8]) -> u24;
+    fn load_u32(src: &[u8]) -> u32;
+
     fn store<O: BitStore, U: Integral>(dest: &mut BitSlice<O>, value: U);
+    fn store_u16(dest: &mut [u8], value: u16);
+    fn store_u24(dest: &mut [u8], value: u24);
+    fn store_u32(dest: &mut [u8], value: u32);
 }
 
 pub struct BigEndian {}
@@ -22,8 +29,46 @@ impl ByteOrder for BigEndian {
         src.load_be()
     }
 
+    fn load_u16(src: &[u8]) -> u16 {
+        u16::from_be_bytes(src.try_into().unwrap())
+    }
+
+    fn load_u24(src: &[u8]) -> u24 {
+        let mut value = 0u32;
+        value |= src[0] as u32;
+        value <<= 8;
+        value |= src[1] as u32;
+        value <<= 8;
+        value |= src[2] as u32;
+
+        u24::new(value)
+    }
+
+    fn load_u32(src: &[u8]) -> u32 {
+        u32::from_be_bytes(src.try_into().unwrap())
+    }
+
     fn store<O: BitStore, U: Integral>(dest: &mut BitSlice<O>, value: U) {
         dest.store_be(value);
+    }
+
+    fn store_u16(dest: &mut [u8], value: u16) {
+        dest[0] = (value >> 8) as u8;
+        dest[1] = value as u8;
+    }
+
+    fn store_u24(dest: &mut [u8], value: u24) {
+        let value: u32 = value.into();
+        dest[0] = (value >> 16) as u8;
+        dest[1] = (value >> 8) as u8;
+        dest[2] = value as u8;
+    }
+
+    fn store_u32(dest: &mut [u8], value: u32) {
+        dest[0] = (value >> 24) as u8;
+        dest[1] = (value >> 16) as u8;
+        dest[2] = (value >> 8) as u8;
+        dest[3] = value as u8;
     }
 }
 
@@ -32,8 +77,45 @@ impl ByteOrder for LittleEndian {
         src.load_le()
     }
 
+    fn load_u16(src: &[u8]) -> u16 {
+        u16::from_le_bytes(src.try_into().unwrap())
+    }
+
+    fn load_u24(src: &[u8]) -> u24 {
+        let mut value = 0u32;
+        value |= src[2] as u32;
+        value <<= 8;
+        value |= src[1] as u32;
+        value <<= 8;
+        value |= src[0] as u32;
+        u24::new(value)
+    }
+
+    fn load_u32(src: &[u8]) -> u32 {
+        u32::from_le_bytes(src.try_into().unwrap())
+    }
+
     fn store<O: BitStore, U: Integral>(dest: &mut BitSlice<O>, value: U) {
         dest.store_le(value)
+    }
+
+    fn store_u16(dest: &mut [u8], value: u16) {
+        dest[0] = value as u8;
+        dest[1] = (value >> 8) as u8;
+    }
+
+    fn store_u24(dest: &mut [u8], value: u24) {
+        let value: u32 = value.into();
+        dest[0] = value as u8;
+        dest[1] = (value >> 8) as u8;
+        dest[2] = (value >> 16) as u8;
+    }
+
+    fn store_u32(dest: &mut [u8], value: u32) {
+        dest[0] = value as u8;
+        dest[1] = (value >> 8) as u8;
+        dest[2] = (value >> 16) as u8;
+        dest[3] = (value >> 24) as u8;
     }
 }