This change optimizes reading and writing byte-aligned values (u8, u16, u24, u32) when the buffer is byte-aligned as well.  

bench before:

```
Timer precision: 36 ns
put_get_ux              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ get_ux_byte_aligned  276 ns        │ 382.6 ns      │ 307.7 ns      │ 311.8 ns      │ 100     │ 1000000
╰─ put_ux_byte_aligned  174.9 ns      │ 224.7 ns      │ 185.5 ns      │ 188.2 ns      │ 100     │ 1000000
```


after:

```
Timer precision: 43 ns
put_get_ux              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ get_ux_byte_aligned  67.45 ns      │ 96.33 ns      │ 76.11 ns      │ 76.61 ns      │ 100     │ 1000000
╰─ put_ux_byte_aligned  69.12 ns      │ 98.73 ns      │ 76.47 ns      │ 77.76 ns      │ 100     │ 1000000
```

Note that even though we check for byte alignment, due to `chain` we can't assume the bytes we're writing to are contiguous, so even with this change we have copy the data into bytes and then write them via `try_copy_to_slice_bytes`/`try_put_slice_bytes` so the reads/writes will play nicely if the underlying buffer is chained.